[ch #161404822]Removed unnecessary files, refactored tests and views
