+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\__init__.py":[1,2,3,4,5,6,7,8,10,11,12,13],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\__init__.py":[1,2,3],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\__init__.py":[1],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\views\\product_views.py":[1,2,3,4,5,6,10,11,12,14,15,16,17,18,19,20,21,22,23,24,25,26,27,29,30,31,32,33,34,35,36,37,39,42,43,44,45,46,47,48,49,51,52,53,54,55,56,57,58,59,60,61,62,63,65,66,67,68,69,72,74,75,78,79,80,81,82,84],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\models.py":[1,2,4,5,6,7,8,9,10,12,13,14,15,16,17,18,19,20,21,23,24,25,26,29,30,32,33,34,35,37,38,39,40,41,42,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,72,73,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,94,96,97,98,99,100,101,102,103,104,105,106,107,108,109,110,112],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\views\\sale_views.py":[1,2,3,4,5,6,10,11,12,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,31,32,33,34,35,36,37,38,39,41,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,68,69,70,71,73,75,76,79,80,81,82,83,86],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\views\\user_view.py":[1,2,3,4,5,6,7,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,33,34,36,38,41,43,44,45,46,47,48,49,50,51,52,53,54,56,57,60,61,64,67,68,69,70,71,72,73,74,79,80,83,84,85,86,89,90,93,94,95,96,97,100,101,103],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\tests\\__init__.py":[1],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\tests\\test_basecase.py":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\tests\\test_product.py":[1,2,3,4,7,8,10,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,28,30,31,32,33,34,35,36,37,38,39,40,41,43,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,63,64,65,66,67,68,69,70,71,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,107,108,109,110,111,112,114,116,117,118,119,120,121,123,126,127,128,129,130,131,132,134,135,136,139],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\tests\\test_sales.py":[1,2,3,4,7,8,10,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,106,107,108,109,110,111,113,114,115,117,120,121,122,124,126,127,128,130,132,133,134,135,136,137,139,141,142,145],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\tests\\test_user.py":[1,2,3,4,7,8,10,12,13,14,15,16,17,18,20,22,23,24,25,26,27,29,31,32,33,34,35,36,38,40,41,42,43,44,45,47,49,50,51,52,53,54,55,56,57,58,59,60,61,63,66,67,70,71,72,73,74,75,76,78,79,80,81,82,83,85,86,87,88,90,92,93,95,96,97,98,100,102,103,105,106,107,109,110,111,112,114,116,119],"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\utils.py":[]}}
+!coverage.py: This is a private format, don't read it directly!{"lines": {"C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\__init__.py": [1, 2, 3, 4, 5, 6, 7, 8, 10, 11, 12, 13], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\__init__.py": [1, 2, 3], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\__init__.py": [1], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\views\\product_views.py": [1, 2, 3, 4, 5, 6, 10, 11, 12, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 29, 30, 31, 32, 33, 34, 35, 36, 37, 39, 42, 43, 44, 45, 46, 47, 48, 49, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 65, 66, 67, 68, 69, 72, 74, 75, 78, 79, 80, 81, 82, 84], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\models.py": [1, 2, 4, 5, 6, 7, 8, 9, 10, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 23, 24, 25, 26, 29, 30, 32, 33, 34, 35, 37, 38, 39, 40, 41, 42, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 72, 73, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 94, 96, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 112], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\views\\sale_views.py": [1, 2, 3, 4, 5, 6, 10, 11, 12, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 31, 32, 33, 34, 35, 36, 37, 38, 39, 41, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 68, 69, 70, 71, 73, 75, 76, 79, 80, 81, 82, 83, 86], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\views\\user_view.py": [1, 2, 3, 4, 5, 6, 7, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 33, 34, 36, 38, 41, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 56, 57, 60, 61, 64, 67, 68, 69, 70, 71, 72, 73, 74, 79, 80, 83, 84, 85, 86, 89, 90, 93, 94, 95, 96, 97, 100, 101, 103], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\tests\\__init__.py": [1], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\tests\\test_basecase.py": [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\tests\\test_product.py": [1, 2, 3, 4, 7, 8, 10, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 28, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 43, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 63, 64, 65, 66, 67, 68, 69, 70, 71, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100, 101, 102, 103, 104, 105, 107, 108, 109, 110, 111, 112, 114, 116, 117, 118, 119, 120, 121, 123, 126, 127, 128, 129, 130, 131, 132, 134, 135, 136, 139], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\tests\\test_sales.py": [1, 2, 3, 4, 7, 8, 10, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 113, 114, 115, 117, 120, 121, 122, 124, 126, 127, 128, 130, 132, 133, 134, 135, 136, 137, 139, 141, 142, 145], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\tests\\test_user.py": [1, 2, 3, 4, 7, 8, 10, 12, 13, 14, 15, 16, 17, 18, 20, 22, 23, 24, 25, 26, 27, 29, 31, 32, 33, 34, 35, 36, 38, 40, 41, 42, 43, 44, 45, 47, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 63, 66, 67, 70, 71, 72, 73, 74, 75, 76, 78, 79, 80, 81, 82, 83, 85, 86, 87, 88, 90, 92, 93, 95, 96, 97, 98, 100, 102, 103, 105, 106, 107, 109, 110, 111, 112, 114, 116, 119], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\utils.py": [], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\models\\product_models.py": [], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\models\\sale_models.py": [], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\models\\user_models.py": [], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\app\\api\\v1\\models\\__init__.py": [], "C:\\Users\\FRED\\Desktop\\stores\\StoreManager-API\\run.py": []}}
