+++ app/tests/test_basecase.py
@@ -18,25 +18,29 @@ class TestSetUp(unittest.TestCase):
         self.token = self.data['token']
         self.app.post("/api/v1/auth/register", data=json.dumps(self.unknown),content_type="application/json")
         self.missing_email = dict(email="", password="testpass")
-        self.product = dict(name="Shoe Polish", description="Kiwi Shoe Polish", price="50", quantity=30)
-        self.new_product = dict(name="Kiatu mzuri", description="Ni kiatu tu", price="1200",
+        self.product = dict(name="Shoe Polish", description="Kiwi Shoe Polish", price=1200, quantity=30)
+        self.new_product = dict(name="Kiatu mzuri", description="Ni kiatu tu", price=1200,
                                 quantity=30)
         self.empty_product = dict(name="", description="", price="", quantity=0)
-        self.missing_prod_quantity = dict(name="Shoe", description="Ni kiatu tu", price="1200",
+        self.missing_prod_quantity = dict(name="Shoe", description="Ni kiatu tu", price=1200,
                                           quantity=0)
+        self.quantity_string = dict(name="Shoe", description="Ni kiatu tu", price=1200,
+                                    quantity="50")
+
         self.missing_prod_description = dict(name="Kiatu mzuri", description="", price="1200",
                                              quantity=10)
-        self.missing_prod_price = dict(name="Kiatu mzuri", description="", price="",
+        self.missing_prod_price = dict(name="Kiatu mzuri", description="Kiatu tu", price=0,
                                        quantity=50)
-        self.new_product = dict(name="Kiatu mzuri", description="Ni kiatu tu", price="1200",
+        self.prod_price_string = dict(name="Kiatu mzuri", description="kiatu", price="100",
+                                      quantity=50)
+        self.new_product = dict(name="Kiatu mzuri", description="Ni kiatu tu", price=1200,
                                 quantity=40)
-        self.sale = dict(name="Shoes", description="All kinds of shoes", quantity=50, total="10000", prod_id=1)
-        self.new_sale = dict(name="Spoons", description="Ni vijiko jamani", quantity=120, total="3000", prod_id=2)
-        self.empty_sale = dict(name="", description="", quantity=0, total="", prod_id=3)
-        self.missing_sale_quantity = dict(name="Spoons", description="Ni vijiko jamani", quantity=0, total="3000",
-                                          prod_id=4)
-        self.missing_sale_description = dict(name="Spoons", description="", quantity=120, total="3000", prod_id=5)
-        self.missing_sale_total = dict(name="Spoons", description="Ni vijiko jamani", quantity=120, total="", prod_id=6)
+        self.sale = dict(prod_id=1, quantity=20)
+        self.new_sale = dict(prod_id=1, quantity=10)
+        self.empty_sale = dict(quantity=0, prod_id=0)
+        self.zero_sale_quantity = dict(quantity=0, prod_id=1)
+        self.sale_quantity_string = dict(quantity="", prod_id=1)
+        self.sale_prod_id_string = dict(quantity=20, prod_id="")
         self.wrong_email_format = dict(email="1234.xxx", password="testpass")
         self.invalid_email = dict(emaile="testuser.gmail.com", password="invalid")
         self.invalid_password = dict(email="testuser3@gmail.com", password="t")
