+++ app/tests/test_sales.py
@@ -1,7 +1,7 @@
 import json
 import unittest
 from app.tests.test_basecase import TestSetUp
-from app.api.v1.models import Sales
+from app.api.v1.models.sale_models import Sales
 
 
 class TestSalesModel(TestSetUp):
@@ -9,16 +9,25 @@ class TestSalesModel(TestSetUp):
 
     def test_sale_creation(self):
         """Tests whether our API can create a sale record"""
+        self.app.post('/api/v1/products',
+                      data=json.dumps(dict(name="Good shoes", description="Good shoes", price=1200,
+                                quantity=30)),
+                      content_type="application/json",
+                      headers={"x-access-token": self.token})
         response = self.app.post('/api/v1/sales',
-                                 data=json.dumps(self.new_sale),
+                                 data=json.dumps(dict(prod_id=1, quantity=5)),
                                  content_type="application/json",
                                  headers={"x-access-token": self.token})
         self.assertEqual(response.status_code, 201)
         response_msg = json.loads(response.data.decode("UTF-8"))
-        self.assertIn("Sale", response_msg["Message"])
+        self.assertIn("updated", response_msg["Message"])
 
     def test_sale_access_with_invalid_token(self):
         """Raise unauthorized error when invalid token is used to access a sale"""
+        self.app.post('/api/v1/products',
+                      data=json.dumps(self.new_product),
+                      content_type="application/json",
+                      headers={"x-access-token": self.token})
         response = self.app.post("/api/v1/sales",
                                  data=json.dumps(self.sale),
                                  content_type="application/json",
@@ -27,7 +36,12 @@ class TestSalesModel(TestSetUp):
 
     def test_find_sale_by_id(self):
         """Tests whether our API can find a sale record by sale id"""
-        self.app.post('/api/v1/sales', data=json.dumps(self.sale),
+        self.app.post('/api/v1/products',
+                      data=json.dumps(self.new_product),
+                      content_type="application/json",
+                      headers={"x-access-token": self.token})
+        self.app.post('/api/v1/sales',
+                      data=json.dumps(self.new_sale),
                       content_type="application/json",
                       headers={"x-access-token": self.token})
         resp = self.app.get('/api/v1/sales/1')
@@ -40,72 +54,29 @@ class TestSalesModel(TestSetUp):
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(resp.content_type, 'application/json')
 
-    def test_missing_sale_name(self):
-        """Test that API should not accept missing sale name"""
+    def test_sale_quantity_is_zero(self):
+        """Test that API should not accept zero sale quantity"""
+        self.app.post('/api/v1/products',
+                      data=json.dumps(self.new_product),
+                      content_type="application/json",
+                      headers={"x-access-token": self.token})
         response = self.app.post("/api/v1/sales",
-                                 data=json.dumps(dict(name="")),
+                                 data=json.dumps(self.zero_sale_quantity),
                                  content_type="application/json",
                                  headers={"x-access-token": self.token})
         self.assertEqual(response.status_code, 400)
         response_msg = json.loads(response.data.decode("UTF-8"))
         self.assertIn("required", response_msg["Message"])
-
-    def test_missing_sale_descr(self):
-        """Test that API should not accept missing sale description"""
-        response = self.app.post("/api/v1/sales",
-                                 data=json.dumps(self.missing_sale_description),
-                                 content_type="application/json",
-                                 headers={"x-access-token": self.token})
-        self.assertEqual(response.status_code, 400)
-        response_msg = json.loads(response.data.decode("UTF-8"))
-        self.assertIn("required", response_msg["Message"])
-
-    def test_missing_quantity(self):
-        """Test that API should not accept missing sale quantity"""
-        response = self.app.post("/api/v1/sales",
-                                 data=json.dumps(self.missing_sale_quantity),
-                                 content_type="application/json",
-                                 headers={"x-access-token": self.token})
-        self.assertEqual(response.status_code, 400)
-        response_msg = json.loads(response.data.decode("UTF-8"))
-        self.assertIn("required", response_msg["Message"])
-
-    def test_missing_total(self):
-        """Test that API should not accept missing sale total"""
-        response = self.app.post("/api/v1/sales",
-                                 data=json.dumps(self.missing_sale_total),
-                                 content_type="application/json",
-                                 headers={"x-access-token": self.token})
-        self.assertEqual(response.status_code, 400)
-        response_msg = json.loads(response.data.decode("UTF-8"))
-        self.assertIn("required", response_msg["Message"])
-
-    def test_sale_update(self):
-        """Tests that the API can update a sale record"""
-        response = self.app.put("/api/v1/sales/1",
-                                data=json.dumps(
-                                    dict(name="Cups", description="Ni vikombe jamani", quantity=20, total="3000")
-                                ),
-                                content_type="application/json",
-                                headers={"x-access-token": self.token})
-        self.assertEqual(response.status_code, 200)
-        response_msg = json.loads(response.data.decode("UTF-8"))
-        self.assertIn("updated", response_msg["Message"])
 
     def test_sale_delete(self):
         """Tests that the API can delete a sale."""
-        self.app.post(
-            '/api/v1/sales',
-            data=json.dumps(
-                dict(
-                    name="Phones",
-                    description="Android and iOS Phones",
-                    quantity=20,
-                    total="50000")),
-            content_type="application/json",
-            headers={
-                "x-access-token": self.token}
-        )
+        self.app.post('/api/v1/products',
+                      data=json.dumps(self.product),
+                      content_type="application/json",
+                      headers={"x-access-token": self.token})
+        self.app.post('/api/v1/sales',
+                      data=json.dumps(self.new_sale),
+                      content_type="application/json", headers={"x-access-token": self.token})
         response = self.app.delete("/api/v1/sales/2",
                                    content_type="application/json",
                                    headers={
@@ -114,19 +85,6 @@ class TestSalesModel(TestSetUp):
         self.assertEqual(response.status_code, 200)
         response_msg = json.loads(response.data.decode("UTF-8"))
         self.assertIn("deleted", response_msg["Message"])
-
-    def test_for_non_duplicates(self):
-        """Tests that API can raise an error if you try to add a
-        product that is in the list"""
-        response = self.app.post("/api/v1/sales",
-                                 data=json.dumps(self.sale),
-                                 content_type="application/json",
-                                 headers={
-                                     "x-access-token": self.token}
-                                 )
-        self.assertEqual(response.status_code, 400)
-        response_msg = json.loads(response.data.decode("UTF-8"))
-        self.assertIn("already exists", response_msg["Message"])
 
     def test_invalid_delete_request(self):
         """Tests that the API raises an error when user places an invalid delete request"""
