+++ requirements.txt
@@ -21,6 +21,7 @@ mccabe==0.6.1
 more-itertools==4.3.0
 nose==1.3.7
 pluggy==0.8.0
+psycopg2==2.7.5
 py==1.7.0
 pycodestyle==2.4.0
 PyJWT==1.6.4
